Fixes for MSVC11.

+++ include/maidsafe/common/utils.h
@@ -181,7 +181,9 @@ fs::path GetUserAppDir();
 // uses kCompanyName and kApplicationName
 fs::path GetSystemAppDir();
 
-uint16_t Concurrency();
+// Returns max of (2, hardware_concurrency)
+unsigned int Concurrency();
+
 // Takes a MaidSafe version as an int and returns the string form, e.g. 901
 // returns "v0.09.01".
 std::string GetMaidSafeVersion(int version,
